+++ src/compas_timber/utils/__init__.py
@@ -4,9 +4,4 @@ from .compas_extra import intersection_line_plane
 from .helpers import close
 from .helpers import are_objects_identical
 
-__all__ = [
-    "intersection_line_line_3D",
-    "intersection_line_plane",
-    "close",
-    "are_objects_identical",
-]
+__all__ = ["intersection_line_line_3D", "intersection_line_plane", "close", "are_objects_identical"]